Initial mongodb proxy
This is a mongodb proxy server that works at the the MongoDB protocol
level so in theory should work with any driver.  It supports adding
plugins that are called at various points during a request/reponse
cycle with MongoDB.  Two plugins are built-in that log the equivalent
shell commands that would be executed and a raw logging plugin that
logs very detailed message data.  This is intended to be used during
development to aid in performance tuning and debugging.

Some things that could possibly be implemented using this:

* Record statisics on various queries
* Introduce slow downs to simulate slow queries
* Introduce errors to simulate MongoDB failures
* Identify N+1 query scenarios
* Find queries that don't use indexes
* ...